Separate the client and server packages
