+++ build.gradle
@@ -27,6 +27,7 @@ dependencies {
     compileOnly 'org.projectlombok:lombok'
     runtimeOnly 'com.h2database:h2'
     annotationProcessor 'org.projectlombok:lombok'
+    compile group: 'com.vaadin', name: 'vaadin', version: '13.0.8'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testCompile group: 'org.spockframework', name: 'spock-core', version: '1.3-groovy-2.5'
     testCompile group: 'org.spockframework', name: 'spock-spring', version: '1.3-groovy-2.5'
